+++ GHEtool/VariableClasses/FlowRateData/_FlowRateData.py
@@ -5,13 +5,13 @@ from abc import ABC
 class _FlowRateData(ABC):
 
     @abc.abstractmethod
-    def vfr(self, *args):
+    def vfr(self, **kwargs):
         """
         This function returns the volume flow rate [l/s].
 
         Parameters
         ----------
-        args
+        kwargs
 
         Returns
         -------
@@ -19,13 +19,13 @@ class _FlowRateData(ABC):
         """
 
     @abc.abstractmethod
-    def mfr(self, *args):
+    def mfr(self, **kwargs):
         """
         This function returns the mass flow rate [kg/s].
 
         Parameters
         ----------
-        args
+        kwargs
 
         Returns
         -------