+++ GHEtool/test/unit-tests/test_borehole.py
@@ -4,20 +4,22 @@ import pygfunction as gt
 import numpy as np
 import pytest
 
-from GHEtool import FluidData, DoubleUTube, SingleUTube, MultipleUTube
+from GHEtool import FluidData, DoubleUTube, SingleUTube, MultipleUTube, ConstantFluidData, ConstantFlowRate, \
+    TemperatureDependentFluidData
 from GHEtool.VariableClasses import Borehole
 
-fluid_data = FluidData(0.2, 0.568, 998, 4180, 1e-3)
+fluid_data = ConstantFluidData(0.568, 998, 4180, 1e-3)
+flow_data = ConstantFlowRate(mfr=0.2)
 pipe_data = DoubleUTube(1, 0.015, 0.02, 0.4, 0.05)
 
 
 def test_fluid_data_without_pipe():
     borehole = Borehole()
-    assert borehole.fluid_data == FluidData()
+    assert borehole.fluid_data is None
     borehole.fluid_data = fluid_data
     assert borehole.fluid_data == fluid_data
     del borehole.fluid_data
-    assert borehole.fluid_data == FluidData()
+    assert borehole.fluid_data is None
 
 
 def test_fluid_data_with_pipe_first_added():
@@ -25,6 +27,8 @@ def test_fluid_data_with_pipe_first_added():
     borehole.pipe_data = pipe_data
     assert borehole.pipe_data.R_f == 0
     borehole.fluid_data = fluid_data
+    assert borehole.pipe_data.R_f == 0
+    borehole.pipe_data.calculate_resistances(fluid_data, flow_data)
     assert np.isclose(0.01663038005086118, borehole.pipe_data.R_f)
 
 
@@ -32,18 +36,21 @@ def test_fluid_data_with_pipe_later_added():
     borehole = Borehole()
     borehole.pipe_data = pipe_data
     borehole.fluid_data = fluid_data
+    assert borehole.pipe_data.R_f == 0
+    borehole.pipe_data.calculate_resistances(fluid_data, flow_data)
     assert np.isclose(0.01663038005086118, borehole.pipe_data.R_f)
 
 
 def test_pipe_data():
     borehole = Borehole()
-    assert borehole.pipe_data == MultipleUTube()
-    assert borehole.pipe_data.R_p == 0
+    assert borehole.pipe_data is None
     borehole.pipe_data = pipe_data
     assert borehole.pipe_data == pipe_data
+    assert borehole.pipe_data.R_p == 0
+    borehole.pipe_data.calculate_resistances(fluid_data, flow_data)
     assert np.isclose(0.11446505967405429, borehole.pipe_data.R_p)
     del borehole.pipe_data
-    assert borehole.pipe_data == MultipleUTube()
+    assert borehole.pipe_data is None
 
 
 def test_equivalent_loading():
@@ -65,9 +72,15 @@ def test_equivalent_loading():
 def test_calculate_Rb():
     borehole = Borehole()
     borehole.pipe_data = MultipleUTube(1, 0.015, 0.02, 0.4, 0.05, 2)
-    borehole.fluid_data = FluidData(0.2, 0.568, 998, 4180, 1e-3)
+    borehole.fluid_data = ConstantFluidData(0.568, 998, 4180, 1e-3)
+    borehole.flow_rate_data = ConstantFlowRate(mfr=0.2)
 
     assert np.isclose(borehole.calculate_Rb(100, 1, 0.075, 3), 0.09483159131195469)
+
+    borehole.fluid_data = TemperatureDependentFluidData('MPG', 25)
+    with pytest.raises(TypeError):
+        np.isclose(borehole.calculate_Rb(100, 1, 0.075, 3), 0.09483159131195469)
+    assert np.isclose(borehole.calculate_Rb(100, 1, 0.075, 3, temperature=20), 0.13807578911314267)
 
 
 def test_Rb():
@@ -77,6 +90,7 @@ def test_Rb():
     # set pipe and fluid data
     borehole.fluid_data = fluid_data
     borehole.pipe_data = pipe_data
+    borehole.flow_rate_data = flow_data
 
     assert np.isclose(borehole.calculate_Rb(100, 1, 0.075, 3), 0.09483159131195469)
 
@@ -87,6 +101,7 @@ def test_Rb():
     # set pipe and fluid data
     borehole.fluid_data = fluid_data
     borehole.pipe_data = pipe_data
+    borehole.flow_rate_data = flow_data
 
     assert np.isclose(borehole.calculate_Rb(100, 1, 0.075, 3), 0.09483159131195469)
     del borehole.pipe_data
@@ -122,7 +137,7 @@ def test_Rb_values():
 
 
 def test_assign_by_initiation():
-    borehole = Borehole(fluid_data, pipe_data)
+    borehole = Borehole(fluid_data, pipe_data, flow_data)
     assert np.isclose(borehole.calculate_Rb(100, 1, 0.075, 3), 0.09483159131195469)
     assert not borehole.use_constant_Rb
 
@@ -130,23 +145,24 @@ def test_assign_by_initiation():
 def test_repr_():
     borehole = Borehole()
     borehole.pipe_data = MultipleUTube(1, 0.015, 0.02, 0.4, 0.05, 2)
-    borehole.fluid_data = FluidData(0.2, 0.568, 998, 4180, 1e-3)
+    borehole.fluid_data = ConstantFluidData(0.568, 998, 4180, 1e-3)
+    borehole.flow_rate_data = ConstantFlowRate(vfr=0.2)
 
-    assert 'Borehole parameters:\n' \
-           'Fluid parameters\n' \
-           '\tThermal conductivity of the fluid [W/(m·K)]: 0.568\n' \
-           '\tDensity of the fluid [kg/m³]: 998.000\n' \
-           '\tThermal capacity of the fluid [J/(kg·K)]: 4180.000\n' \
-           '\tDynamic viscosity [Pa·s]: 0.001\n' \
-           '\tMass flow rate [kg/s] : 0.2\n' \
-           'U tube\n' \
-           '\tNumber of U tubes [-]: 2\n' \
-           '\tPipe wall thickness [mm]: 5.0\n' \
-           '\tPipe diameter [mm]: 40.0\n' \
-           '\tDistance from pipe to borehole center [mm]: 50\n' \
-           '\tGrout conductivity [W/(m·K)]: 1\n' \
-           '\tPipe conductivity [W/(m·K)]: 0.4\n' \
-           '\tPipe roughness [mm]: 0.001' == borehole.__repr__()
+    assert {'fluid': {'Pr': 7.359154929577465,
+                      'cp [J/(kg·K)]': 4180,
+                      'freezing_point [°C]': None,
+                      'k_f [W/(m·K)]': 0.568,
+                      'mu [Pa·s]': 0.001,
+                      'nu [m²/s]': 1.002004008016032e-06,
+                      'rho [kg/m³]': 998},
+            'pipe': {'diameter [mm]': 4.0,
+                     'epsilon [mm]': 0.001,
+                     'k_g [W/(m·K)]': 1,
+                     'k_p [W/(m·K)]': 0.4,
+                     'nb_of_tubes': 2,
+                     'spacing [mm]': 50.0,
+                     'thickness [mm]': 5.0,
+                     'type': 'U'}} == borehole.__repr__()
 
     borehole = Borehole()
-    assert 'Borehole effective thermal resistance [(m·K)/W]: 0.12' == borehole.__repr__()
+    assert {'Rb': 0.12} == borehole.__repr__()